brute_force_zip/fixed writing zip issue

+++ brute_force_zip/main.c
@@ -29,14 +29,14 @@ static size_t write_to_buffer(void *contents, size_t size, size_t nmemb,
 
 static size_t write_to_file(void *ptr, size_t size, size_t nmemb,
                             void *stream) {
-  return fwrite(ptr, size, nmemb, stream);
+  return fwrite(ptr, size, nmemb, (FILE *)stream);
 }
 
 int main(void) {
   CURL *curl = NULL;
   FILE *fptr = NULL;
+  cJSON *root = NULL;
   char *zip_url = NULL;
-  memory_t response = {0};
   int rc = EXIT_FAILURE;
 
   const char *token = getenv("HACKATTIC_TOKEN");
@@ -61,6 +61,8 @@ int main(void) {
            token);
 
   /* JSON request  */
+  memory_t response = {0};
+
   curl_easy_setopt(curl, CURLOPT_URL, url);
   curl_easy_setopt(curl, CURLOPT_FOLLOWLOCATION, 1L);
   curl_easy_setopt(curl, CURLOPT_WRITEFUNCTION, write_to_buffer);
@@ -70,7 +72,7 @@ int main(void) {
     goto cleanup;
   }
 
-  cJSON *root = cJSON_Parse(response.data);
+  root = cJSON_Parse(response.data);
   if (!root) {
     goto cleanup;
   }
@@ -83,10 +85,11 @@ int main(void) {
 
   zip_url = strdup(zip->valuestring);
   cJSON_Delete(root);
+  root = NULL;
   free(response.data);
   response.data = NULL;
 
-  /* file download */
+  /* writing the zip file */
   fptr = fopen("protected.zip", "wb");
   if (!fptr) {
     goto cleanup;
@@ -99,6 +102,9 @@ int main(void) {
   if (curl_easy_perform(curl) != CURLE_OK) {
     goto cleanup;
   }
+
+  fclose(fptr);
+  fptr = NULL;
 
   /* run fcrackzip */
   fptr = popen("fcrackzip protected.zip -c a1 -l 4-6 -v -u 2>&1", "r");
@@ -127,33 +133,45 @@ int main(void) {
   }
 
   /* unzip protected.zip */
-  // char unzip_cmd[512];
-  // snprintf(unzip_cmd, sizeof(unzip_cmd), "7z x protected.zip -p'%s'",
-  // password); system(unzip_cmd);
-  //
-  // /* read the secret.txt */
-  // fptr = fopen("secret.txt", "r");
-  // char secret[100];
-  // fgets(secret, sizeof(secret), fptr);
-  //
-  // /* send the secret to the solution endpoint */
-  // char solution_url[512];
-  // char json_data[512];
-  // snprintf(json_data, sizeof(json_data), "{\"secret\":\"%s\"}", secret);
-  // snprintf(
-  //     solution_url, sizeof(solution_url),
-  //     "https://hackattic.com/challenges/brute_force_zip/solve?access_token=%s",
-  //     token);
-  //
-  // curl_easy_setopt(curl, CURLOPT_URL, solution_url);
-  // curl_easy_setopt(curl, CURLOPT_POSTFIELDS, json_data);
-  // curl_easy_setopt(curl, CURLOPT_POSTFIELDSIZE, sizeof(json_data));
-  //
-  // if (curl_easy_perform(curl) != CURLE_OK) {
-  //   goto cleanup;
-  // }
-  //
+  char unzip_cmd[512];
+  snprintf(unzip_cmd, sizeof(unzip_cmd), "7z x protected.zip -p'%s'", password);
+  system(unzip_cmd);
+
+  /* read the secret.txt */
+  fptr = fopen("secret.txt", "r");
+  char secret[100];
+  fgets(secret, sizeof(secret), fptr);
+
+  /* send the secret to the solution endpoint */
+  memory_t post_response = {0};
+
+  char solution_url[512];
+  char json_data[512];
+  snprintf(json_data, sizeof(json_data), "{\"secret\":\"%s\"}", secret);
+  snprintf(
+      solution_url, sizeof(solution_url),
+      "https://hackattic.com/challenges/brute_force_zip/solve?access_token=%s",
+      token);
+  printf("json=%s\n", json_data);
+
+  curl_easy_setopt(curl, CURLOPT_URL, solution_url);
+  curl_easy_setopt(curl, CURLOPT_POST, 1L);
+  curl_easy_setopt(curl, CURLOPT_POSTFIELDS, json_data);
+  curl_easy_setopt(curl, CURLOPT_POSTFIELDSIZE, strlen(json_data));
+  curl_easy_setopt(curl, CURLOPT_WRITEFUNCTION, write_to_buffer);
+  curl_easy_setopt(curl, CURLOPT_WRITEDATA, &post_response);
+  curl_easy_setopt(curl, CURLOPT_ACCEPT_ENCODING, "");
+  curl_easy_setopt(curl, CURLOPT_FOLLOWLOCATION, 1L);
+
+  if (curl_easy_perform(curl) != CURLE_OK) {
+    goto cleanup;
+  }
+
+  printf("Hackattic response: %s\n", post_response.data);
+  printf("Cleaning data files\n");
+
   system("rm *.txt");
+  system("rm protected*");
 
   rc = EXIT_SUCCESS;
